RESKC-575: Fixing attachment removal

+++ coeus-impl/src/main/java/org/kuali/coeus/propdev/impl/person/attachment/ProposalPersonBiography.java
@@ -457,7 +457,7 @@ public class ProposalPersonBiography extends KcPersistableBusinessObjectBase imp
 
     @PostRemove
     public void removeData() {
-        if (getPersonnelAttachment().getFileDataId() != null) {
+        if (getPersonnelAttachment() != null && getPersonnelAttachment().getFileDataId() != null) {
             getKcAttachmentDao().removeData(getPersonnelAttachment().getFileDataId());
         }
     }
